component Main is smashed into 2 components: MainScreen and MainOffers
