+++ src/components/main-offers/main-offers.e2e.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import {mount} from 'enzyme';
-import Main from './main.jsx';
+import MainOffers from './main-offers.jsx';
 
 const EMPTY_HANDLER = () => {};
 const CITY = {
@@ -34,12 +34,12 @@ const APARTMENTS = [
   }
 ];
 
-describe(`<Main /> e2e suite`, () => {
-  it(`<Main /> Apartment title is pressed`, () => {
+describe(`<MainOffers /> e2e suite`, () => {
+  it(`<MainOffers /> Apartment title is pressed`, () => {
     const onApartmentTitleHandler = jest.fn();
 
-    const mainWrapper = mount(
-        <Main
+    const mainOffersWrapper = mount(
+        <MainOffers
           activeCity={CITY}
           cityList={[CITY]}
           apartmentList={APARTMENTS}
@@ -48,7 +48,7 @@ describe(`<Main /> e2e suite`, () => {
         />
     );
 
-    const apartmentButtons = mainWrapper.find(`.place-card__name > a`);
+    const apartmentButtons = mainOffersWrapper.find(`.place-card__name > a`);
     const eventMock = {preventDefault() {}};
     apartmentButtons.at(0).simulate(`click`, eventMock);
